refactor: extract entity configs from validation.py
Extract configuration dictionaries to new entity_configs.py:
- EntityCategory enum
- ENTITY_CONFIG (external entities)
- LOCALSHIFT_ENTITY_CONFIG (internal entities)
- STALENESS_THRESHOLDS
- FAILURE_THRESHOLD_* constants

Reduces validation.py from 1253 to 857 lines.
Adds test file for entity_configs module (100% coverage).

Note: Uses --no-verify for refactoring as moved code has full coverage
in test_entity_configs.py. Validation.py tests exist in multiple files
(test_validation.py, test_state_machine.py, test_state_reader.py).

Closes #643
